+++ ContabilidadDbContext/ContabilidadDbContext.cs
@@ -8,7 +8,7 @@ namespace DbContextLibrary
     public class ContabilidadDbContext : DbContext
     {
         private const string DbName = "Contabilidad"; // TODO: Define database name
-        
+
         private static IConfigurationRoot? _configuration;
 
         public DbSet<Cliente> Clientes { get; set; }
@@ -47,6 +47,11 @@ namespace DbContextLibrary
             modelBuilder.Entity<Cliente>()
                 .Property(c => c.TipoDocumento)
                 .HasConversion<EnumToStringConverter<TipoDocumento>>();
+
+            modelBuilder.Entity<Cliente>().HasData(
+                new Cliente { Nombre = "Imporcom", NumeroDocumento = "123-456-7890", TipoDocumento = TipoDocumento.Nit },
+                new Cliente { Nombre = "Andres' Programmers SAS", NumeroDocumento = "111-222-33-44" }
+            );
         }
 
 
@@ -60,6 +65,5 @@ namespace DbContextLibrary
         //     _dbContextOptionsBuilder = new DbContextOptionsBuilder<ContabilidadDbContext>();
         //     _dbContextOptionsBuilder.UseSqlServer(_configuration.GetConnectionString(DbName));
         // }
-
     }
 }
