+++ ManejoContable/MainWindow.xaml
@@ -18,13 +18,13 @@
                 <!-- TODO: Implementar Funcionalidad de los MenuItems  -->
                 <MenuItem Header="Menu Principal" />
 
-                <MenuItem Header="Salir" Click="ExitMenutItem_OnClick" />
+                <MenuItem Header="Salir" Click="ExitMenuItem_OnClick" />
             </MenuItem>
             <MenuItem Header="Clientes">
                 <MenuItem Name="ViewClientsMenuItem"
                           Header="Ver Clientes" 
                           Click="ViewClientsMenuItem_OnClick" />
-                <MenuItem Header="Agregar" Click="MenuItem_OnClick" />
+                <MenuItem Header="Agregar" Click="AddClientMenuItem_OnClick" />
                 <MenuItem Header="Eliminar" Click="MenuItem_OnClick" />
                 <MenuItem Header="Buscar" Click="MenuItem_OnClick" />
