+++ ManejoContable/ManejoContable.csproj
@@ -39,7 +39,6 @@
 
 
 	<ItemGroup>
-	  <Folder Include="UserControls\ClientUserControls\" />
 	  <Folder Include="ViewModel\" />
 	</ItemGroup>
