+++ ManejoContable/UserControls/ClientUserControls/EditUserInformationControl.xaml
@@ -1,0 +1,200 @@
+﻿<UserControl x:Class="ManejoContable.UserControls.ClientUserControls.EditUserInformationControl"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ManejoContable.UserControls.ClientUserControls"
+             xmlns:modelEntities="clr-namespace:ModelEntities;assembly=ModelEntities"
+             xmlns:system="clr-namespace:System;assembly=System.Runtime"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+
+    <UserControl.Resources>
+        <modelEntities:Cliente x:Key="Cliente" />
+
+        <Style TargetType="Label">
+            <Setter Property="FontWeight" Value="Bold" />
+            <Setter Property="VerticalAlignment" Value="Center" />
+        </Style>
+        <Style TargetType="TextBox">
+
+        </Style>
+
+        <ObjectDataProvider x:Key="TipoDocumentoSource"
+                            ObjectType="{x:Type system:Enum}"
+                            MethodName="GetValues">
+            <ObjectDataProvider.MethodParameters>
+                <x:Type TypeName="modelEntities:TipoDocumento"/>
+            </ObjectDataProvider.MethodParameters>
+
+        </ObjectDataProvider>
+    </UserControl.Resources>
+
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="Auto" />
+
+        </Grid.ColumnDefinitions>
+
+
+        <Label Content="Numero Factura"
+               Grid.Row="0"
+               Grid.Column="0" />
+
+        <Label Content="Nombre Comercial"
+               Grid.Row="0"
+               Grid.Column="2" />
+
+        <Label Content="Tipo Documento"
+               Grid.Row="1"
+               Grid.Column="0" />
+
+        <Label Content="Numero Documento"
+               Grid.Row="1"
+               Grid.Column="2" />
+
+        <Label Content="Tipo Contribuyente"
+               Grid.Row="4"
+               Grid.Column="2" />
+
+
+        <TextBox Grid.Row="0"
+                 Grid.Column="1"
+                 Text="Numero Factura" />
+
+        <ComboBox Name="TipoDocumentoComboBox"
+                  Grid.Row="1"
+                  Grid.Column="1"
+                  Text="[CC, NIT, ...]"
+                  SelectedIndex="0"
+                  ItemsSource="{Binding Source={StaticResource TipoDocumentoSource}}">
+
+        </ComboBox>
+
+        <TextBox Grid.Row="1"
+                 Grid.Column="3"
+                 Text="Num Doc" />
+
+        <ComboBox Grid.Row="4"
+                  Grid.Column="3"
+                  SelectedIndex="0">
+
+            <ComboBoxItem Content="[persona juridica]"></ComboBoxItem>
+            <ComboBoxItem Content="this is 1" />
+            <ComboBoxItem Content="this is 2" />
+            <ComboBoxItem Content="there are 3" />
+            <ComboBoxItem Content="there are 4" />
+            <ComboBoxItem Content="but 5" />
+        </ComboBox>
+
+        <TextBox Grid.Row="0"
+                 Grid.Column="3"
+                 Text="Nombre Client" />
+
+
+        <Label Content="Regimen Contable"
+               Grid.Column="0"
+               Grid.Row="5" />
+
+        <Label Content="Tipo Responsabilidad"
+               Grid.Row="5" Grid.Column="2" />
+
+        <Label Content="Departamento"
+               Grid.Column="0"
+               Grid.Row="2" />
+
+        <Label Content="Municipio"
+               Grid.Column="2"
+               Grid.Row="2" />
+
+        <Label Content="Direccion"
+               Grid.Column="0"
+               Grid.Row="3" />
+
+        <Label Content="Correo"
+               Grid.Column="2"
+               Grid.Row="3" />
+
+        <Label Content="Telefono"
+               Grid.Column="0"
+               Grid.Row="4" />
+
+        <TextBlock Grid.Column="1"
+                   Grid.Row="5"
+                   HorizontalAlignment="Center"
+                   Text="???" />
+
+        <TextBlock Grid.Column="3"
+                   Grid.Row="5"
+                   HorizontalAlignment="Center"
+                   Text="???" />
+
+
+        <ComboBox Grid.Column="1"
+                  Grid.Row="2"
+                  SelectedIndex="0">
+            <ComboBoxItem Content="Departamento" />
+            <ComboBoxItem Content="this is 1" />
+            <ComboBoxItem Content="this is 2" />
+            <ComboBoxItem Content="there are 3" />
+            <ComboBoxItem Content="there are 4" />
+            <ComboBoxItem Content="but 5" />
+        </ComboBox>
+        <ComboBox Grid.Column="3"
+                  Grid.Row="2"
+                  SelectedIndex="0">
+            <ComboBoxItem Content="Municipio" />
+            <ComboBoxItem Content="this is 1" />
+            <ComboBoxItem Content="this is 2" />
+            <ComboBoxItem Content="there are 3" />
+            <ComboBoxItem Content="there are 4" />
+            <ComboBoxItem Content="but 5" />
+        </ComboBox>
+        <TextBox Grid.Row="3"
+                 Grid.Column="1"
+                 Text="Direccion" />
+        <TextBox Grid.Row="4"
+                 Grid.Column="1"
+                 Text="Telefono" />
+
+        <TextBox Grid.Row="3"
+                 Grid.Column="3"
+                 Text="Correo" />
+
+        <StackPanel Grid.Row="6" Grid.Column="0"
+                    Grid.ColumnSpan="4"
+                    HorizontalAlignment="Right"
+                    Orientation="Horizontal">
+            <StackPanel.Resources>
+                <Style TargetType="Button">
+                    <Setter Property="Margin" Value="5, 10" />
+                    <Setter Property="Padding" Value="15, 5" />
+                </Style>
+            </StackPanel.Resources>
+
+            <Button Name="CancelButton"
+                    Content="Cancelar"
+                    Click="ActionButton_OnClick" />
+
+            <Button Name="OkButton"
+                    Content="Aceptar"
+                    Click="ActionButton_OnClick" />
+
+        </StackPanel>
+    </Grid>
+
+</UserControl>
